+++ src/hooks/useTabLayout.ts
@@ -4,64 +4,11 @@ import {
   TabMenuType,
   activeTabType,
   activeTabTypeState,
-  tabMenuTypeState,
 } from "../states/layout/layoutAtom";
-import { menuItems } from "../constants/data";
 
 function useTabMenu() {
-  const [tabMenuState, setTabMenuState] =
-    useRecoilState<TabMenuListType>(tabMenuTypeState);
   const [activeTab, setActiveTab] =
     useRecoilState<activeTabType>(activeTabTypeState);
-
-  const findElement = (link: string) => {
-    let element = null;
-    menuItems.forEach((item) => {
-      if (item.child) {
-        item.child.forEach((i) => {
-          if (i.multi_menu) {
-            i.multi_menu.forEach((m) => {
-              if (m.multiLink === link) {
-                element = m.multiElement;
-              }
-            });
-          } else if (i.childlink === link) {
-            element = i.childElement;
-          }
-        });
-      } else if (item.link === link) {
-        element = item.element;
-      }
-    });
-    return element;
-  };
-
-  const findTitle = (link: string) => {
-    let title: Array<string> = [];
-    if (link) {
-      menuItems.forEach((item) => {
-        if (item.child) {
-          item.child.forEach((i) => {
-            if (i.multi_menu) {
-              i.multi_menu.forEach((m) => {
-                if (m.multiLink === link) {
-                  title.push(item.title);
-                  title.push(i.childtitle ?? "");
-                  title.push(m.multiTitle);
-                }
-              });
-            } else if (i.childlink === link) {
-              title.push(item.title);
-              title.push(i.childtitle ?? "");
-            }
-          });
-        } else if (item.link === link) {
-          title.push(item.title);
-        }
-      });
-    }
-    return title;
-  };
 
   // 탭 열기
   const handleTabOpen = (
@@ -105,13 +52,8 @@ function useTabMenu() {
 
   return {
     activeTab,
-    setActiveTab,
-    tabMenuState,
-    setTabMenuState,
     handleTabOpen,
     handleTabClose,
-    findElement,
-    findTitle,
   };
 }
