+++ src/layout/Layout.tsx
@@ -1,25 +1,27 @@
 import { useEffect, useState } from "react";
 import Breadcrumbs from "../components/breadcrumbs/Breadcrumbs";
 import TabMenu from "../components/tabMenu/TabMenu";
-import useTabMenu from "../hooks/useTabMenu";
+import useTabMenu from "../hooks/useTabLayout";
 import Error404Page from "../pages/error/Error404Page";
-import { TabMenuListType } from "../states/layout/layoutAtom";
+import { TabMenuListType, tabMenuTypeState } from "../states/layout/layoutAtom";
 import Footer from "../components/footer/Footer";
 import Header from "../components/header/Header";
 import Sidebar from "../components/sidebar/Sidebar";
+import tabLayoutUtil from "../utils/tabLayoutUtil";
+import { useRecoilState } from "recoil";
 
 function Layout() {
   const [tabMenu, setTabMenu] = useState<TabMenuListType>([]);
-
-  const { activeTab, tabMenuState, setTabMenuState, findElement } =
-    useTabMenu();
+  const [tabMenuState, setTabMenuState] =
+    useRecoilState<TabMenuListType>(tabMenuTypeState);
+  const { activeTab } = useTabMenu();
 
   useEffect(() => {
     setTabMenu(
       tabMenuState.map(({ name, href }) => ({
         name,
         href,
-        component: findElement(href),
+        component: tabLayoutUtil.findElement(href),
       }))
     );
   }, []);
@@ -63,35 +65,6 @@ function Layout() {
           <Footer />
         </div>
       </main>
-      {/* <main className="flex flex-grow hidden w-full">
-        <Sidebar tabMenu={tabMenu} setTabMenu={setTabMenu} />
-        <div className="flex flex-col w-full h-full">
-          <TabMenu tabMenu={tabMenu} setTabMenu={setTabMenu} />
-          <div className="flex flex-col grow">
-            {tabMenu.map((tab) => (
-              <div
-                key={tab.href}
-                className="flex flex-col h-full"
-                style={{
-                  display: activeTab === tab.href ? "block" : "none",
-                }}
-              >
-                {tab.component ? (
-                  <section className="flex flex-col h-full p-3 grow">
-                    <Breadcrumbs activeTab={activeTab} />
-                    <div className="bg-white rounded-[10px] p-3 overflow-y-auto grow">
-                      <tab.component />
-                    </div>
-                  </section>
-                ) : (
-                  <Error404Page />
-                )}
-              </div>
-            ))}
-          </div>
-          <Footer />
-        </div>
-      </main> */}
     </div>
   );
 }
